+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add GPG signatures to all winesapOS packages and the database metadata.
 - Change the secondary installed Linux kernel from Linux LTS 5.15 to Linux T2 (currently based on Linux 6.4).
 - Change Replay Sorcery to be installed via the 'replay-sorcery-git' package.
+- Add fingerprint scanning support.
 
 ## 3.3.0 - The Major Upgrade Update